refactor(whatsrust): extract group and chat queries
